Move test to callbacks_test
